We can significantly speed up route matching by not using `URLPattern` and opting for a custom constructed `RegExp` instead. This is roughly 150-180x faster and I confirm that in live traces this greatly reduces CPU time.

```sh
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------
pattern       160.12 µs/iter       6,245.2 (151.88 µs … 712.62 µs) 161.04 µs 191.46 µs 221.21 µs
regex         889.44 ns/iter   1,124,305.9 (870.65 ns … 965.37 ns) 892.66 ns 965.37 ns 965.37 ns
```

I've added an additional optimization that first checks if we're dealing with a pattern or an exact path match in the first place. If it's an exact match we can skip parsing it to a pattern entirely, which is even faster.